+++ frontend-web/My-Syncra/src/App.jsx
@@ -12,6 +12,7 @@ import Chat from './Components/Sections/Chat'
 import Network from './Components/Sections/Network'
 import Profile from './Components/Sections/Profile'
 import Portfolio from './Components/Sections/Portfolio' // <-- Add this import
+import Job from './Components/Sections/Job'; // Import your Job component
 
 function App() {
   const [isLoaded, setIsLoaded] = useState(false)
@@ -74,6 +75,14 @@ function App() {
               element={
                 <PrivateRoute>
                   <Portfolio />
+                </PrivateRoute>
+              }
+            />
+            <Route
+              path="/jobs"
+              element={
+                <PrivateRoute>
+                  <Job />
                 </PrivateRoute>
               }
             />
