+++ frontend-web/My-Syncra/src/Components/NavBar.jsx
