--- frontend-web/My-Syncra/src/Components/Sections/Portfolio.css
@@ -1,281 +1,0 @@
-/* Root layout */
-.portfolio-root {
-  display: flex;
-  min-height: 100vh;
-  background: #f5f6fa;
-  padding-top: 64px; /* Height of NavBar */
-}
-
-/* Sidebar */
-.portfolio-sidebar {
-  width: 270px;
-  min-width: 220px;
-  background: #fff;
-  padding: 24px 0 24px 0;
-  border-radius: 16px;
-  margin: 16px 0 16px 16px;
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-  box-shadow: 0 2px 16px rgba(0,0,0,0.04);
-  height: fit-content;
-}
-
-.portfolio-promo {
-  background: #1a6ed8;
-  border-radius: 20px;
-  padding: 24px 12px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.portfolio-promo-content h2 {
-  color: #fff;
-  font-size: 1.3rem;
-  font-weight: 700;
-  margin-bottom: 24px;
-  line-height: 1.2;
-}
-
-.portfolio-promo-content button {
-  background: #fff;
-  color: #1a6ed8;
-  border: none;
-  border-radius: 12px;
-  padding: 10px 28px;
-  font-weight: 600;
-  font-size: 1rem;
-  cursor: pointer;
-  transition: background 0.2s;
-}
-
-.portfolio-promo-content button:hover {
-  background: #e3e9f7;
-}
-
-.portfolio-filters {
-  padding: 0 18px;
-}
-
-.portfolio-filters h3 {
-  font-size: 1.1rem;
-  font-weight: 700;
-  margin-bottom: 12px;
-}
-
-.portfolio-filter-group {
-  margin-bottom: 18px;
-}
-
-.portfolio-filter-title {
-  font-size: 0.98rem;
-  font-weight: 600;
-  margin-bottom: 6px;
-}
-
-.portfolio-checkbox {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  font-size: 0.97rem;
-  margin-bottom: 4px;
-  cursor: pointer;
-}
-
-.portfolio-checkbox input[type="checkbox"] {
-  accent-color: #1a6ed8;
-}
-
-/* Main content */
-.portfolio-main {
-  flex: 1;
-  padding: 32px 32px 32px 32px;
-  min-width: 0;
-}
-
-.portfolio-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 18px;
-  flex-wrap: wrap;
-}
-
-.portfolio-header h1 {
-  font-size: 2rem;
-  font-weight: 700;
-  margin: 0;
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.portfolio-count {
-  background: #e3e9f7;
-  color: #1a6ed8;
-  border-radius: 12px;
-  font-size: 1rem;
-  padding: 2px 10px;
-  font-weight: 600;
-}
-
-.portfolio-sort {
-  font-size: 1rem;
-  color: #888;
-}
-
-.portfolio-sort-link {
-  color: #1a6ed8;
-  text-decoration: underline;
-  margin-left: 4px;
-  font-weight: 500;
-  cursor: pointer;
-}
-
-.portfolio-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(260px, 1fr));
-  gap: 24px;
-}
-
-.portfolio-card {
-  background: linear-gradient(135deg, #b3d0f7 0%, #e3e9f7 100%);
-  border-radius: 16px;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.08);
-  padding: 18px 18px 16px 18px;
-  display: flex;
-  flex-direction: column;
-  border: 1.5px solid #dbeafe;
-  min-width: 0;
-  position: relative;
-}
-
-.portfolio-card-date {
-  position: absolute;
-  top: 14px;
-  left: 14px;
-  background: #fff;
-  color: #1a6ed8;
-  font-size: 0.85rem;
-  font-weight: 600;
-  border-radius: 8px;
-  padding: 2px 10px;
-  box-shadow: 0 1px 4px rgba(0,0,0,0.04);
-}
-
-.portfolio-card-company {
-  margin-top: 32px;
-  font-size: 1rem;
-  font-weight: 600;
-  color: #222;
-}
-
-.portfolio-card-title {
-  font-size: 1.2rem;
-  font-weight: 700;
-  color: #1a365d;
-  margin-bottom: 10px;
-}
-
-.portfolio-card-tags {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-  margin-bottom: 18px;
-}
-
-.portfolio-tag {
-  background: #fff;
-  color: #1a6ed8;
-  border-radius: 8px;
-  font-size: 0.85rem;
-  padding: 3px 10px;
-  font-weight: 500;
-  box-shadow: 0 1px 4px rgba(0,0,0,0.04);
-}
-
-.portfolio-card-bottom {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-top: auto;
-}
-
-.portfolio-card-rate {
-  font-size: 1.15rem;
-  font-weight: 700;
-  color: #222;
-}
-
-.portfolio-card-details {
-  background: #1a6ed8;
-  color: #fff;
-  border: none;
-  border-radius: 10px;
-  padding: 8px 22px;
-  font-weight: 600;
-  font-size: 1rem;
-  cursor: pointer;
-  transition: background 0.2s;
-}
-
-.portfolio-card-details:hover {
-  background: #155bb5;
-}
-
-/* Responsive styles */
-@media (max-width: 1100px) {
-  .portfolio-root {
-    flex-direction: column;
-    padding-top: 64px;
-  }
-  .portfolio-sidebar {
-    margin: 16px auto 0 auto;
-    width: 95%;
-    min-width: 0;
-    flex-direction: row;
-    gap: 32px;
-    justify-content: space-between;
-    align-items: flex-start;
-  }
-  .portfolio-main {
-    padding: 24px 8px 24px 8px;
-  }
-}
-
-@media (max-width: 800px) {
-  .portfolio-root {
-    flex-direction: column;
-    padding-top: 56px;
-  }
-  .portfolio-sidebar {
-    width: 100%;
-    margin: 12px 0 0 0;
-    border-radius: 0;
-    flex-direction: column;
-    align-items: stretch;
-    gap: 18px;
-  }
-  .portfolio-main {
-    padding: 12px 2vw;
-  }
-  .portfolio-header h1 {
-    font-size: 1.2rem;
-  }
-}
-
-@media (max-width: 600px) {
-  .portfolio-main {
-    padding: 6px 1vw;
-  }
-  .portfolio-header {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 8px;
-  }
-  .portfolio-grid {
-    grid-template-columns: 1fr;
-    gap: 14px;
-  }
-}
